FOGL-625 moved microservice management and service registry out of core
