+++ src/python/foglamp/microservice_management/service_registry/service_registry.py
@@ -8,7 +8,7 @@
 
 import time
 from aiohttp import web
-from foglamp.core.service_registry.instance import Service
+from foglamp.microservice_management.service_registry.instance import Service
 
 __author__ = "Amarendra Kumar Sinha, Praveen Garg"
 __copyright__ = "Copyright (c) 2017 OSIsoft, LLC"
